Bump Rasa to version 3.6.2 and update all dependencies, scripts, README.md
Also apply formatting/changes of config.yml that Rasa automatically applies
if `assistant_id` is missing. Rasa now automatically adds a new value
`assistant_id` (whatever this is good for ...) and therefore reformatting
of config.yml happens.

As this assistant_id should be unique, it's not added to the CI

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>